Move MaskStates classes to its own file independent to Animate package
