+++ korge/src/commonMain/kotlin/com/soywiz/korge/render/RenderContext.kt
@@ -42,6 +42,8 @@ class RenderContext constructor(
 
     var debugAnnotateView: View? = null
 
+    var stencilIndex: Int = 0
+
     /** Allows to draw quads, sprites and nine patches using a precomputed global matrix or raw vertices */
 	val batch by lazy { BatchBuilder2D(this) }
